+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -1,8 +1,16 @@
-package nextstep.subway.application;
+package nextstep.subway.line.application;
 
 import nextstep.subway.consts.ErrorMessage;
-import nextstep.subway.domain.*;
-import nextstep.subway.dto.*;
+import nextstep.subway.section.dto.*;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.LineUpdateRequest;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
+import nextstep.subway.station.dto.StationResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
