+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,7 +1,10 @@
-package nextstep.subway.ui;
+package nextstep.subway.line.ui;
 
-import nextstep.subway.application.LineService;
-import nextstep.subway.dto.*;
+import nextstep.subway.line.application.LineService;
+import nextstep.subway.section.dto.*;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.LineUpdateRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
