+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -1,4 +1,7 @@
-package nextstep.subway.domain;
+package nextstep.subway.section.domain;
+
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
