+++ src/test/java/nextstep/subway/domain/SectionsTest.java
@@ -1,6 +1,8 @@
 package nextstep.subway.domain;
 
-import org.assertj.core.api.Assertions;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.domain.Sections;
+import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
