+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -3,10 +3,10 @@ package nextstep.subway.section;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.dto.SectionRequest;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.station.dto.StationResponse;
 import nextstep.subway.line.LineTestMethods;
 import nextstep.subway.station.StationTestMethods;
 import org.junit.jupiter.api.BeforeEach;
